+++ packages/components/radio/src/index.ts
@@ -2,6 +2,7 @@ export { Radio, useRadio } from "./radio"
 export type { RadioProps, UseRadioProps, UseRadioReturn } from "./radio"
 export { RadioGroup, useRadioGroup } from "./radio-group"
 export type {
+  RadioItem,
   RadioGroupProps,
   UseRadioGroupProps,
   UseRadioGroupReturn,