+++ stories/components/forms/multi-autocomplete.stories.tsx
@@ -9,7 +9,7 @@ import {
   MultiAutocomplete,
   AutocompleteOptionGroup,
   AutocompleteOption,
-  UIOption,
+  AutocompleteItem,
   Tag,
   FormControl,
   VStack,
@@ -28,11 +28,11 @@ const meta: Meta<typeof MultiAutocomplete> = {
 export default meta
 
 export const basic: Story = () => {
-  const options: UIOption[] = [
+  const items: AutocompleteItem[] = [
     { label: "ベジータ", value: "ベジータ" },
     {
       label: "地球人",
-      value: [
+      items: [
         { label: "孫悟空", value: "孫悟空" },
         { label: "孫悟飯", value: "孫悟飯" },
         { label: "クリリン", value: "クリリン" },
@@ -40,7 +40,7 @@ export const basic: Story = () => {
     },
     {
       label: "フリーザ軍",
-      value: [
+      items: [
         { label: "フリーザ", value: "フリーザ" },
         { label: "ギニュー", value: "ギニュー" },
         { label: "リクーム", value: "リクーム" },
@@ -76,7 +76,7 @@ export const basic: Story = () => {
         </AutocompleteOptionGroup>
       </MultiAutocomplete>
 
-      <MultiAutocomplete placeholder="キャラクターを選択" options={options} />
+      <MultiAutocomplete placeholder="キャラクターを選択" items={items} />
     </>
   )
 }
@@ -129,12 +129,12 @@ export const withEmptyMessage: Story = () => {
   )
 }
 
-export const withCreateOption: Story = () => {
-  const options: UIOption[] = [
+export const withAllowCreate: Story = () => {
+  const items: AutocompleteItem[] = [
     { label: "ベジータ", value: "ベジータ" },
     {
       label: "地球人",
-      value: [
+      items: [
         { label: "孫悟空", value: "孫悟空" },
         { label: "孫悟飯", value: "孫悟飯" },
         { label: "クリリン", value: "クリリン" },
@@ -142,7 +142,7 @@ export const withCreateOption: Story = () => {
     },
     {
       label: "フリーザ軍",
-      value: [
+      items: [
         { label: "フリーザ", value: "フリーザ" },
         { label: "ギニュー", value: "ギニュー" },
         { label: "リクーム", value: "リクーム" },
@@ -156,21 +156,21 @@ export const withCreateOption: Story = () => {
   return (
     <MultiAutocomplete
       placeholder="キャラクターを選択"
-      options={options}
-      createOption
-      onCreate={(newOption, newOptions) =>
-        console.log("created option", newOption, "new options", newOptions)
+      items={items}
+      allowCreate
+      onCreate={(newItem, newItems) =>
+        console.log("created item", newItem, "new items", newItems)
       }
     />
   )
 }
 
-export const withCreateOrder: Story = () => {
-  const options: UIOption[] = [
+export const withInsertPositionItem: Story = () => {
+  const items: AutocompleteItem[] = [
     { label: "ベジータ", value: "ベジータ" },
     {
       label: "地球人",
-      value: [
+      items: [
         { label: "孫悟空", value: "孫悟空" },
         { label: "孫悟飯", value: "孫悟飯" },
         { label: "クリリン", value: "クリリン" },
@@ -178,7 +178,7 @@ export const withCreateOrder: Story = () => {
     },
     {
       label: "フリーザ軍",
-      value: [
+      items: [
         { label: "フリーザ", value: "フリーザ" },
         { label: "ギニュー", value: "ギニュー" },
         { label: "リクーム", value: "リクーム" },
@@ -193,27 +193,27 @@ export const withCreateOrder: Story = () => {
     <>
       <MultiAutocomplete
         placeholder="キャラクターを選択"
-        options={options}
-        createOption
-        insertPositionOnCreate="first"
+        items={items}
+        allowCreate
+        insertPositionItem="first"
       />
       <MultiAutocomplete
         placeholder="キャラクターを選択"
-        options={options}
-        createOption
-        insertPositionOnCreate="last"
+        items={items}
+        allowCreate
+        insertPositionItem="last"
       />
       <MultiAutocomplete
         placeholder="キャラクターを選択"
-        options={options}
-        createOption
-        insertPositionOnCreate="地球人"
+        items={items}
+        allowCreate
+        insertPositionItem="地球人"
       />
       <MultiAutocomplete
         placeholder="キャラクターを選択"
-        options={options}
-        createOption
-        insertPositionOnCreate={["フリーザ軍", "last"]}
+        items={items}
+        allowCreate
+        insertPositionItem={["フリーザ軍", "last"]}
       />
     </>
   )
@@ -582,11 +582,11 @@ export const reactHookForm: Story = () => {
     autocomplete3: string[]
   }
 
-  const options: UIOption[] = [
+  const items: AutocompleteItem[] = [
     { label: "ベジータ", value: "ベジータ" },
     {
       label: "地球人",
-      value: [
+      items: [
         { label: "孫悟空", value: "孫悟空" },
         { label: "孫悟飯", value: "孫悟飯" },
         { label: "クリリン", value: "クリリン" },
@@ -594,7 +594,7 @@ export const reactHookForm: Story = () => {
     },
     {
       label: "フリーザ軍",
-      value: [
+      items: [
         { label: "フリーザ", value: "フリーザ" },
         { label: "ギニュー", value: "ギニュー" },
         { label: "リクーム", value: "リクーム" },
@@ -688,7 +688,7 @@ export const reactHookForm: Story = () => {
             <MultiAutocomplete
               placeholder="キャラクターを選択"
               {...field}
-              options={options}
+              items={items}
             />
           )}
         />
@@ -714,11 +714,11 @@ export const reactHookFormWithDefaultValue: Story = () => {
     autocomplete3: ["リクーム"],
   }
 
-  const options: UIOption[] = [
+  const items: AutocompleteItem[] = [
     { label: "ベジータ", value: "ベジータ" },
     {
       label: "地球人",
-      value: [
+      items: [
         { label: "孫悟空", value: "孫悟空" },
         { label: "孫悟飯", value: "孫悟飯" },
         { label: "クリリン", value: "クリリン" },
@@ -726,7 +726,7 @@ export const reactHookFormWithDefaultValue: Story = () => {
     },
     {
       label: "フリーザ軍",
-      value: [
+      items: [
         { label: "フリーザ", value: "フリーザ" },
         { label: "ギニュー", value: "ギニュー" },
         { label: "リクーム", value: "リクーム" },
@@ -820,7 +820,7 @@ export const reactHookFormWithDefaultValue: Story = () => {
             <MultiAutocomplete
               placeholder="キャラクターを選択"
               {...field}
-              options={options}
+              items={items}
             />
           )}
         />